Fix stale value, add network delay correction
The retry timeout was set to 1sec, which occasioanlly caused a retry
that would produce two responses. Since only one response is processed,
the second response would get queued up and lead to a stale reading on
the next update. Setting the rety to 10sec is a quick fix. But ideally,
we'd use the origin timestamp in the response packet to detect stale
responses. Unfortunately, the origin timestamp doesn't seem to get set
as specified in the SNTP spec. Maybe something in the request isn't
quite right?

Also adds a network delay correction by computing the round-trip delay
and assuming the response delay is half that.